Converted much of wx node viewer to use libzinc API. Added new API to get query node derivatives, versions and time, get field component names and current time from time notifier. Still depends on internal field manager and FE_region change callbacks. https://tracker.physiomeproject.org/show_bug.cgi?id=3410

+++ source/computed_field/field_cache.hpp
@@ -275,7 +275,7 @@ public:
 		}
 		else
 		{
-			for (int i = valueCaches.size(); i <= cacheIndex; ++i)
+			for (int i = static_cast<int>(valueCaches.size()); i <= cacheIndex; ++i)
 				valueCaches.push_back(0);
 		}
 		valueCaches[cacheIndex] = valueCache;
